% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.R
\name{init_cleaning_log}
\alias{init_cleaning_log}
\title{Initiate an empty cleaning log}
\usage{
init_cleaning_log(dir, filename = "cleaning_logbook.csv",
  extra_columns = NULL)
}
\arguments{
\item{dir}{The directory name of the location where it needs to be saved.}

\item{filename}{Name of the logbook. Default is cleaning_logbook.csv. Needs to have a csv extention.}

\item{extra_columns}{Must be a list with the names of the objects being the names of the columns in the cleaning log
and the content of the list being the names of the dataset columns associated with it (see example). This is not required.}
}
\value{
Nothing, it just writes the empty logbook.
}
\description{
Initiate an empty cleaning log
}
\examples{
dir <- getwd()
extra_columns <- list(population_group = "population_group", governorate = "governorate_mcna")
\dontrun{
init.cleaning.log(dir, extra_columns = extra_columns)
}
}
