% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcna2019-data.R
\docType{data}
\name{mcna2019}
\alias{mcna2019}
\title{Extract of Multi Cluster Needs Assessment 2019 for Iraq}
\format{An object of class \code{"data.frame"};}
\source{
\href{REACH resource center}{REACH Resource Center}
}
\usage{
data(mcna2019)
}
\description{
A dataset...
}
\examples{
data(mcna2019)
}
\references{
Multi Cluster Needs Assessment Iraq, REACH (2019)
(\href{REACH resource center}{PubMed})
}
\keyword{datasets}
